backend/requirements.txt
fastapi
python-socketio[asyncio_server]
httpx
uvicorn
